# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: WIB - Second Brain API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: priority
          schema:
            type: string
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectInput"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project details
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectInput"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks
      parameters:
        - in: query
          name: projectId
          schema:
            type: integer
        - in: query
          name: priority
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: today
          schema:
            type: boolean
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskInput"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskInput"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    Project:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        clientName:
          type: string
        type:
          type: string
        status:
          type: string
        priority:
          type: string
        description:
          type: string
        aiSummary:
          type: string
        progress:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, type, status, priority, createdAt]
    ProjectDetail:
      allOf:
        - $ref: "#/components/schemas/Project"
        - type: object
          properties:
            tasks:
              type: array
              items:
                $ref: "#/components/schemas/Task"
            people:
              type: array
              items:
                $ref: "#/components/schemas/Person"
            files:
              type: array
              items:
                $ref: "#/components/schemas/FileRecord"
            timeline:
              type: array
              items:
                $ref: "#/components/schemas/TimelineEvent"
          required: [tasks, people, files, timeline]
    CreateProjectInput:
      type: object
      properties:
        name:
          type: string
        clientName:
          type: string
        type:
          type: string
        status:
          type: string
        priority:
          type: string
        description:
          type: string
      required: [name, type]
    UpdateProjectInput:
      type: object
      properties:
        name:
          type: string
        clientName:
          type: string
        type:
          type: string
        status:
          type: string
        priority:
          type: string
        description:
          type: string
        progress:
          type: integer
    Person:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
        organization:
          type: string
        createdAt:
          type: string
      required: [id, name, role, createdAt]
...
    Task:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        title:
          type: string
        description:
          type: string
        priority:
          type: string
        status:
          type: string
        dueDate:
          type: string
        createdAt:
          type: string
        projectName:
          type: string
      required: [id, title, priority, status, createdAt]
    CreateTaskInput:
      type: object
      properties:
        projectId:
          type: integer
        title:
          type: string
        description:
          type: string
        priority:
          type: string
        status:
          type: string
        dueDate:
          type: string
      required: [title, priority]
    UpdateTaskInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        priority:
          type: string
        status:
          type: string
        dueDate:
          type: string
...
    FileRecord:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        name:
          type: string
        fileType:
          type: string
        url:
          type: string
        size:
          type: integer
        aiSummary:
          type: string
        createdAt:
          type: string
      required: [id, name, fileType, createdAt]
    TimelineEvent:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        type:
          type: string
        title:
          type: string
        description:
          type: string
        projectName:
          type: string
        createdAt:
          type: string
      required: [id, type, title, createdAt]
...