# lib/api-zod/src/generated/types/createProjectInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WIB - Second Brain API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateProjectInput {
  name: string;
  clientName?: string;
  type: string;
  status?: string;
  priority?: string;
  description?: string;
}